Move GetTemplates to internal/drapi package
